refactor: 비동기 스레드 풀 설정 분리
