+++ space-d/src/main/java/com/dnd/spaced/global/config/AppConfig.java
@@ -14,8 +14,6 @@ import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
 import java.time.Clock;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
-import java.util.concurrent.Executor;
-import java.util.concurrent.ThreadPoolExecutor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
 import org.springframework.context.annotation.Bean;
@@ -24,7 +22,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.ResourceHttpMessageConverter;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -46,7 +43,6 @@ public class AppConfig implements WebMvcConfigurer {
     private final GradedAnswerPageableArgumentResolver gradedAnswerPageableArgumentResolver;
     private final WordPageableArgumentResolver wordPageableArgumentResolver;
 
-
     @Bean
     public Clock clock() {
         return Clock.systemDefaultZone();
@@ -63,48 +59,6 @@ public class AppConfig implements WebMvcConfigurer {
             builder.simpleDateFormat(DATE_TIME_FORMAT);
             builder.serializers(new LocalDateTimeSerializer(DateTimeFormatter.ofPattern(DATE_TIME_FORMAT)));
         };
-    }
-
-    @Bean
-    public Executor asyncWordViewCountExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-
-        int availableProcessors = Runtime.getRuntime().availableProcessors();
-        int corePoolSize = Math.max(1, availableProcessors / 6);
-        int maxPoolSize = Math.max(2, corePoolSize * 2);
-
-        executor.setCorePoolSize(corePoolSize);
-        executor.setMaxPoolSize(maxPoolSize);
-        executor.setQueueCapacity(10);
-        executor.setKeepAliveSeconds(180);
-        executor.setAllowCoreThreadTimeOut(true);
-        executor.setAwaitTerminationSeconds(20);
-        executor.setRejectedExecutionHandler(new ThreadPoolExecutor.AbortPolicy());
-        executor.setThreadNamePrefix("Async-Word-ViewCount-");
-        executor.initialize();
-
-        return executor;
-    }
-
-    @Bean
-    public Executor asyncCommentLikeCountExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-
-        int availableProcessors = Runtime.getRuntime().availableProcessors();
-        int corePoolSize = Math.max(1, availableProcessors / 6);
-        int maxPoolSize = Math.max(2, corePoolSize * 2);
-
-        executor.setCorePoolSize(corePoolSize);
-        executor.setMaxPoolSize(maxPoolSize);
-        executor.setQueueCapacity(10);
-        executor.setKeepAliveSeconds(180);
-        executor.setAllowCoreThreadTimeOut(true);
-        executor.setAwaitTerminationSeconds(20);
-        executor.setRejectedExecutionHandler(new ThreadPoolExecutor.AbortPolicy());
-        executor.setThreadNamePrefix("Async-Comment-LikeCount-");
-        executor.initialize();
-
-        return executor;
     }
 
     @Override
